divide files for extension
